requirements.txt
streamlit
Pillow
numpy
torch
torchvision
ultralytics

